% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sars2phylo.R
\name{constructHaploTable}
\alias{constructHaploTable}
\title{construct a table of haplotypes by month}
\usage{
constructHaploTable(SS)
}
\arguments{
\item{SS}{dataframe of sample set, including SS$jmonth and SS$Sgeno as columns}
}
\value{
table of haplotypes by month
}
\description{
construct a table of haplotypes by month
}
\examples{
MA_haplo <- constructHaploTable(ss);
}
