% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sars2phylo.R
\name{convertMonth}
\alias{convertMonth}
\title{convert month to sequential month (jmonth) and week to sequential week (jweek):}
\usage{
convertMonth(SS)
}
\arguments{
\item{SS}{dataframe of sample set, including SS$week and SS$month as columns}
}
\value{
dataframe augment with jweek and jmonth columns
}
\description{
convert month to sequential month (jmonth) and week to sequential week (jweek):
}
\examples{
ss <- convertMonth(ss);
}
