% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sars2phylo.R
\name{countByInterval}
\alias{countByInterval}
\title{create count of a given lineage by given time interval}
\usage{
countByInterval(SS, lineage, timeInt)
}
\arguments{
\item{SS}{dataframe of sample set, with one-hot encoding of each lineage as a column}

\item{timeInt}{time interval, usually weekly ("epidate")}
}
\value{
datafram of absolute counts by lineage by week
}
\description{
create count of a given lineage by given time interval
}
\examples{
countByIntrval(ss, "B.1.1.7", "epidate");
}
