% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sars2phylo.R
\name{createSgenoTable}
\alias{createSgenoTable}
\title{create a table of genotypes in a lineage}
\usage{
createSgenoTable(SS)
}
\arguments{
\item{SS}{dataframe of sample set, must have columns seqName and Sgeno}
}
\value{
vector of frequencies
}
\description{
create a table of genotypes in a lineage
}
\examples{
linTable <- createSgenotTable(ss);
}
