% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sars2phylo.R
\name{encodeGenotype}
\alias{encodeGenotype}
\title{encode genotypes as columns using one-hot encoding}
\usage{
encodeGenotype(SS, spike = FALSE, smallversion = TRUE, N = 100)
}
\arguments{
\item{SS}{dataframe of sample set}

\item{spike}{restrict analysis only to Spike gene}

\item{smallversion}{if true, limit number of genotypes to N}

\item{N}{limits number of genotypes encoded}
}
\value{
dataframe augmented with columns as one-hot encoded genotypes
}
\description{
encode genotypes as columns using one-hot encoding
}
\examples{
ss <- encodeGenotype(ss);
}
