% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sars2phylo.R
\name{encodeLineage}
\alias{encodeLineage}
\title{one-hot encoding lineage as binary outcome, as augmented columns}
\usage{
encodeLineage(SS)
}
\arguments{
\item{SS}{dataframe of sample set}
}
\value{
dataframe augmented with one-hot encoded vectors of each lineage
}
\description{
one-hot encoding lineage as binary outcome, as augmented columns
}
\examples{
ss <- endodeLineage(ss);
}
