% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sars2phylo.R
\name{extractS}
\alias{extractS}
\title{extract S haplotypes from a list of mutations (e.g. called by nextclade)}
\usage{
extractS(mutlist)
}
\arguments{
\item{mutlist}{comma-separated vector}
}
\value{
S haplotype
}
\description{
extract S haplotypes from a list of mutations (e.g. called by nextclade)
}
