% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sars2phylo.R
\name{logRegFit}
\alias{logRegFit}
\title{fit logistic regression and return coefficients and p values}
\usage{
logRegFit(SS, START)
}
\arguments{
\item{SS}{sample set}

\item{START}{column at which the one-hot encoding of the lineages begins}
}
\value{
list length (ncol(SS) - START) of model fits
}
\description{
fit logistic regression and return coefficients and p values
}
\examples{
modFit <- logRegFit(ss, 50)
}
