% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sars2phylo.R
\name{plotCounts}
\alias{plotCounts}
\title{plot absolute counts by timeInt}
\usage{
plotCounts(SS, lineage, timeInt, smoother = FALSE)
}
\arguments{
\item{SS}{dataframe of sample set, with one-hot encoding of each lineage as a column}

\item{lineage}{character string of particular lineage}

\item{timeInt}{time interval, usually weekly ("epidate")}

\item{smoother}{display loess smoother}
}
\value{
creates a plot of proportion of lineage by time interval
}
\description{
plot absolute counts by timeInt
}
\examples{
plotCounts(ss, "B.1.1.7", "epidate");
}
