% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sars2phylo.R
\name{plotLineage}
\alias{plotLineage}
\title{Plot lineage as successes/failures over time, and fit logistic regression as smoother}
\usage{
plotLineage(
  SS,
  LIN,
  bystate = FALSE,
  xlower = "2020-05-01",
  xupper = "2021-05-01"
)
}
\arguments{
\item{SS}{dataframe of sample set, with one-hot encoding of each lineage as a column}

\item{LIN}{character string of particular lineage}

\item{bystate}{subset by state; SS must have state as a column}

\item{xlower}{lower bound of x range}

\item{xupper}{upper bound of x range, will extrapolate the logistic regression smoother}
}
\value{
creates a plot of successes (samples belong to lineage) vs failures (not lineage) over time
}
\description{
Plot lineage as successes/failures over time, and fit logistic regression as smoother
}
\examples{
plotLineage(ss, "B.1.1.7", bystate = TRUE);
}
