% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sars2phylo.R
\name{plotProp}
\alias{plotProp}
\title{plot proportion by time interval with 95% CI}
\usage{
plotProp(SS, lineage, timeInt, smoother = FALSE, logreg = FALSE)
}
\description{
plot proportion by time interval with 95% CI
}
