% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sars2phylo.R
\name{plotTop}
\alias{plotTop}
\title{plot top scoring lineages}
\usage{
plotTop(mn_sum, K)
}
\arguments{
\item{mn_sum}{summary table of multinomial regression}

\item{K}{the number of top lineages to plot}
}
\value{
creates a plot (using plotProp) of top lineages by week, with CIs and best fit
}
\description{
plot top scoring lineages
}
\examples{
plotTop(multinom_sum, 6)
}
