% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sars2phylo.R
\name{plotVarWeekly}
\alias{plotVarWeekly}
\title{Create plotting function for lineage with x bins as week, not day}
\usage{
plotVarWeekly(SS, LIN, bystate = FALSE, xlower, xupper)
}
\arguments{
\item{SS}{dataframe of sample set, with one-hot encoding of each lineage as a column}

\item{LIN}{character string of particular lineage}

\item{bystate}{subset by state; SS must have state as a column}

\item{xlower}{lower bound of x range}

\item{xupper}{upper bound of x range, will extrapolate the logistic regression smoother}
}
\value{
creates a plot of successes (samples belong to lineage) vs failures (not lineage) over time
}
\description{
Create plotting function for lineage with x bins as week, not day
}
\examples{
plotVarWeekly(ss, "B.1.1.7", bystate = FALSE, xlower = 25, xupper = 75);
}
