% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sars2phylo.R
\name{plotVOC}
\alias{plotVOC}
\title{plot variants of concern (B.1.1.7, B.1.351, P.1)}
\usage{
plotVOC(SS, timeInt, smoother = FALSE, ggtit)
}
\arguments{
\item{SS}{dataframe of sample set, with one-hot encoding of each lineage as a column}

\item{timeInt}{time interval, usually weekly ("epidate")}

\item{smoother}{display loess smoother}

\item{ggtit}{title for graph}
}
\value{
creates a plot of absolute counts by lineage by week
}
\description{
plot variants of concern (B.1.1.7, B.1.351, P.1)
}
\examples{
plotVOC(ss, "B.1.1.7", "epidate", "VOC by Week");
}
