% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sars2phylo.R
\name{plotWeekly}
\alias{plotWeekly}
\title{plot lineages as successes by week (for accurate uncertainty in logistic regression smoother)}
\usage{
plotWeekly(SS, lineage, timeInt, smoother = FALSE, logreg = FALSE)
}
\arguments{
\item{SS}{dataframe of sample set, with one-hot encoding of each lineage as a column}

\item{lineage}{character string specifying lineage of interest}

\item{timeInt}{time interval, usually weekly ("epidate")}

\item{smoother}{display loess smoother}

\item{logreg}{display logistic regression smoother}
}
\value{
creates a plot of successes by week for a given lineage
}
\description{
plot lineages as successes by week (for accurate uncertainty in logistic regression smoother)
}
\examples{
plotWeekly(ss, "B.1.1.7", "epidate", logreg=TRUE);
}
