% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sars2phylo.R
\name{readNextClade}
\alias{readNextClade}
\title{read and parse nextclade results}
\usage{
readNextClade(fpath, exten)
}
\arguments{
\item{fpath}{string specifying file path of directory containing one or more nextclade csv files}

\item{exten}{prefix for matching nextclade csv files (in the case of multiple) or filename (if single)}
}
\value{
dataframe with merged values
}
\description{
read and parse nextclade results
}
\examples{
ss <- readNextClade("path","nextclade.csv");
}
