% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sars2phylo.R
\name{readSampleSet}
\alias{readSampleSet}
\title{read sample set}
\usage{
readSampleSet(nextclade_csv, ss_metadata)
}
\arguments{
\item{nextclade_csv}{nextclade csv file}

\item{ss_metadata}{gisaid metadata}
}
\value{
dataframe with merged values
}
\description{
read sample set
}
\examples{
ss <- readSampleSet("nextclade.csv", "metadata_2021_03_29.csv.gz");
}
