% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sars2phylo.R
\name{readTerra}
\alias{readTerra}
\title{read terra assemblies tsv}
\usage{
readTerra(terra_csv)
}
\arguments{
\item{terra_csv}{terra assemblies tsv file}
}
\value{
dataframe with merged values
}
\description{
read terra assemblies tsv
}
\examples{
ss <- readTerra("assemblies.tsv");
}
