% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sars2phylo.R
\name{relabel_DNAss}
\alias{relabel_DNAss}
\title{relabel sequences for BEAST}
\usage{
relabel_DNAss(DNA_string_set, ss_metadata)
}
\arguments{
\item{DNA_string_set}{Biostrings DNAstringset}

\item{ss_metadata}{sample set dataframe with a column ss$beast (with | delimiters) and a column ss$seqName}
}
\value{
DNA_string_set with names formatted for a beast analysis
}
\description{
relabel sequences for BEAST
}
\examples{
DNAss <- relabel_DNAss(DNAss);
}
